requirements.txt
python-telegram-bot>=20.0,<21.0
pytesseract>=0.3.10
Pillow>=10.0.0
exifread>=3.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
geopy>=2.4.0
python-dateutil>=2.8.0
